+++ panda/board/main.c
@@ -722,10 +722,10 @@ void tick_handler(void) {
           if (current_safety_mode != SAFETY_NOOUTPUT) {
             set_safety_mode(SAFETY_NOOUTPUT, 0U);
           }
+
           //if (power_save_status != POWER_SAVE_STATUS_ENABLED) {
           //  set_power_save_state(POWER_SAVE_STATUS_ENABLED);
           //}
-		  
 
           // Also disable IR when the heartbeat goes missing
           current_board->set_ir_power(0U);
@@ -758,7 +758,7 @@ void tick_handler(void) {
       ignition_can_cnt += 1U;
 
       // synchronous safety check
-      safety_tick(current_hooks);
+      safety_tick(current_rx_checks);
     }
 
     loop_counter++;