+++ selfdrive/ui/qt/home.cc
@@ -82,19 +82,8 @@ void HomeWindow::mousePressEvent(QMouseEvent* e) {
   }
 
   // Handle sidebar collapsing
-  else if (onroad->isVisible() && (!sidebar->isVisible() || e->x() > sidebar->width())) {
-
-    // TODO: Handle this without exposing pointer to map widget
-    // Hide map first if visible, then hide sidebar
-    if (onroad->map != nullptr && onroad->map->isVisible()) {
-      onroad->map->setVisible(false);
-    } else if (!sidebar->isVisible()) {
-      sidebar->setVisible(true);
-    } else {
-      sidebar->setVisible(false);
-
-      if (onroad->map != nullptr) onroad->map->setVisible(true);
-    }
+  if (onroad->isVisible() && (!sidebar->isVisible() || e->x() > sidebar->width())) {
+    sidebar->setVisible(!sidebar->isVisible() && !onroad->isMapVisible());
   }
 }